requirements.txt
pyHamcrest
allure-behave
allure-python-commons
attrs
behave
behave-webdriver
importlib-metadata
namegenerator
names
parse
parse-type
pip
pluggy
selenium
selenium-utils
setuptools
six
zipp
python-dotenv
requests
ipdb
pandas
pyautogui

